feat(api): separate `api/spawn.js` from `api/deploy.js`
